fixed a migration added some functionlaity to getFile sevice.

+++ database/migrations/1_create_initial_up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE files (
   id int(11) NOT NULL AUTO_INCREMENT,
   s3Path varchar(255) NOT NULL,
-  ttl int(11) NOT NULL,
+  ttl bigint NOT NULL,
   shortUrl varchar(255) NOT NULL,
   created datetime NOT NULL DEFAULT CURRENT_TIMESTAMP,
   PRIMARY KEY (id)
